+++ tests/test_excellib.py
@@ -12,13 +12,10 @@ from pycel.excellib import (
     averageifs,
     ceiling,
     column,
-    concat,
-    concatenate,
     conditional_format_ids,
     count,
     countif,
     countifs,
-    find,
     floor,
     hlookup,
     index,
@@ -27,29 +24,24 @@ from pycel.excellib import (
     isna,
     isnumber,
     istext,
-    left,
     # ::TODO:: finish test cases for remainder of functions
     # linest,
     ln,
     log,
     lookup,
     match,
-    mid,
     mod,
     npv,
     power,
-    right,
     roundup,
     row,
     sumif,
     sumifs,
     sumproduct,
     trunc,
-    value,
     vlookup,
     x_abs,
     x_int,
-    x_len,
     x_round,
     xatan2,
     xmax,
@@ -181,25 +173,6 @@ def test_column(address, expected):
 
 @pytest.mark.parametrize(
     'args, result', (
-        ('a 1 abc'.split(), 'a1abc'),
-        ('a Jan-00 abc'.split(), 'aJan-00abc'),
-        ('a	#DIV/0! abc'.split(), DIV0),
-        ('a	1 #DIV/0!'.split(), DIV0),
-        ('a #NAME? abc'.split(), NAME_ERROR),
-        (('a', True, 'abc'), 'aTRUEabc'),
-        (('a', False, 'abc'), 'aFALSEabc'),
-        (('a', 2, 'abc'), 'a2abc'),
-    )
-)
-def test_concatenate(args, result):
-    assert concat(*args) == result
-    assert concatenate(*args) == result
-    assert concat(args) == result
-    assert concatenate(args) == VALUE_ERROR
-
-
-@pytest.mark.parametrize(
-    'args, result', (
         (((True, 1, 0), (True, 2, 1), (True, 3, 0)), [1, 2]),
         (((False, 1, 0), (True, 2, 1), (True, 3, 0)), [2]),
         (((False, 1, 0), (True, 2, 0), (True, 3, 0)), [2, 3]),
@@ -264,27 +237,6 @@ class TestCountIfs:
     def test_countifs_odd_args_len(self):
         with pytest.raises(AssertionError):
             countifs(((7, 25, 13, 25), ), 25, ((100, 102, 201, 20), ))
-
-
-@pytest.mark.parametrize(
-    'to_find, find_in, result', (
-        (2, 2.5, 1),
-        ('.', 2.5, 2),
-        (5, 2.5, 3),
-        ('2', 2.5, 1),
-        ('.', 2.5, 2),
-        ('5', 2.5, 3),
-        ('2', '2.5', 1),
-        ('.', '2.5', 2),
-        ('T', True, 1),
-        ('U', True, 3),
-        ('u', True, VALUE_ERROR),
-        (DIV0, 'x' + DIV0, DIV0),
-        ('V', DIV0, DIV0),
-    )
-)
-def test_find(to_find, find_in, result):
-    assert find(to_find, find_in) == result
 
 
 @pytest.mark.parametrize(
@@ -539,27 +491,6 @@ def test_istext(value, expected):
 
 
 @pytest.mark.parametrize(
-    'text, num_chars, expected', (
-        ('abcd', 5, 'abcd'),
-        ('abcd', 4, 'abcd'),
-        ('abcd', 3, 'abc'),
-        ('abcd', 2, 'ab'),
-        ('abcd', 1, 'a'),
-        ('abcd', 0, ''),
-
-        (1.234, 3, '1.2'),
-
-        ('abcd', -1, VALUE_ERROR),
-        ('abcd', 'x', VALUE_ERROR),
-        (DIV0, 1, DIV0),
-        ('abcd', NAME_ERROR, NAME_ERROR),
-    )
-)
-def test_left(text, num_chars, expected):
-    assert left(text, num_chars) == expected
-
-
-@pytest.mark.parametrize(
     'expected, value',
     (
         (math.log(5), 5),
@@ -781,37 +712,6 @@ def test_match_crazy_order(
             assert result1 == _match(lookup_value, lookup_array, 1)
 
 
-class TestMid:
-
-    def test_invalid_parameters(self):
-        assert mid(VALUE_ERROR, 2, 2) == VALUE_ERROR
-        assert mid('Romain', VALUE_ERROR, 2) == VALUE_ERROR
-        assert mid('Romain', 2, VALUE_ERROR) == VALUE_ERROR
-        assert mid(DIV0, 2, 2) == DIV0
-        assert mid('Romain', DIV0, 2) == DIV0
-        assert mid('Romain', 2, DIV0) == DIV0
-
-        assert mid('Romain', 'x', 2) == VALUE_ERROR
-        assert mid('Romain', 2, 'x') == VALUE_ERROR
-
-    def test_num_chars_must_be_integer(self):
-        assert 'Ro' == mid('Romain', 1, 2.1)
-
-    def test_start_num_must_be_superior_or_equal_to_1(self):
-        assert VALUE_ERROR == mid('Romain', 0, 3)
-
-    def test_num_chars_must_be_positive(self):
-        assert VALUE_ERROR == mid('Romain', 1, -1)
-
-    def test_from_not_str(self):
-        assert '23' == mid(1234, 2, 2)
-
-    def test_mid(self):
-        assert 'omain' == mid('Romain', 2, 9)
-        assert 'om' == mid('Romain', 2.1, 2)
-        assert 'om' == mid('Romain', 2, 2.1)
-
-
 class TestMod:
 
     def test_first_argument_validity(self):
@@ -876,27 +776,6 @@ def test_power(data, expected):
         assert result == expected
     else:
         assert result == pytest.approx(expected, rel=1e-3)
-
-
-@pytest.mark.parametrize(
-    'text, num_chars, expected', (
-        ('abcd', 5, 'abcd'),
-        ('abcd', 4, 'abcd'),
-        ('abcd', 3, 'bcd'),
-        ('abcd', 2, 'cd'),
-        ('abcd', 1, 'd'),
-        ('abcd', 0, ''),
-
-        (1234.1, 2, '.1'),
-
-        ('abcd', -1, VALUE_ERROR),
-        ('abcd', 'x', VALUE_ERROR),
-        (VALUE_ERROR, 1, VALUE_ERROR),
-        ('abcd', VALUE_ERROR, VALUE_ERROR),
-    )
-)
-def test_right(text, num_chars, expected):
-    assert right(text, num_chars) == expected
 
 
 @pytest.mark.parametrize(
@@ -1023,30 +902,6 @@ def test_trunc(number, num_digits, result):
 
 
 @pytest.mark.parametrize(
-    'param, result', (
-        (0, 0),
-        (2, 2),
-        (2.1, 2.1),
-        (-2.1, -2.1),
-        ('-2.1', -2.1),
-        ('3', 3),
-        ('3.', 3),
-        ('3.0', 3),
-        ('.01', 0.01),
-        ('1E5', 100000),
-        ('X', VALUE_ERROR),
-        ('`1', VALUE_ERROR),
-        (False, VALUE_ERROR),
-        (True, VALUE_ERROR),
-        (NA_ERROR, NA_ERROR),
-        (DIV0, DIV0),
-    )
-)
-def test_value(param, result):
-    assert value(param) == result
-
-
-@pytest.mark.parametrize(
     'lkup, col_idx, result, approx', (
         ('A', 0, VALUE_ERROR, True),
         ('A', 1, 'A', True),
@@ -1102,22 +957,6 @@ def test_xatan2(param1, param2, result):
 )
 def test_x_int(value, expected):
     assert x_int(value) == expected
-
-
-@pytest.mark.parametrize(
-    'param, result', (
-        ('A', 1),
-        ('BB', 2),
-        (3.0, 3),
-        (True, 4),
-        (False, 5),
-        (None, 0),
-        (NA_ERROR, NA_ERROR),
-        (DIV0, DIV0),
-    )
-)
-def test_x_len(param, result):
-    assert x_len(param) == result
 
 
 def test_xmax():
